requirements.txt
flask
python-dotenv
google-genai
anthropic
gunicorn
openai
flake8
